+++ Assets/LDtkUnity/Editor/Builders/LDtkLevelBackgroundBuilder.cs
@@ -8,15 +8,17 @@ namespace LDtkUnity.Editor
         private readonly GameObject _levelTransform;
         private readonly LDtkSortingOrder _layerSortingOrder;
         private readonly Level _level;
-        
+        private readonly Vector2 _worldSpaceSize;
+
         private Texture2D _texture;
 
-        public LDtkLevelBackgroundBuilder(LDtkProjectImporter importer, GameObject levelTransform, LDtkSortingOrder layerSortingOrder, Level level)
+        public LDtkLevelBackgroundBuilder(LDtkProjectImporter importer, GameObject levelTransform, LDtkSortingOrder layerSortingOrder, Level level, Vector2 worldSpaceSize)
         {
             _importer = importer;
             _levelTransform = levelTransform;
             _layerSortingOrder = layerSortingOrder;
             _level = level;
+            _worldSpaceSize = worldSpaceSize;
         }
 
 
@@ -25,15 +27,33 @@ namespace LDtkUnity.Editor
         /// </returns>
         public void BuildBackground()
         {
-            //if no path defined, then no background was set
+            BuildBackgroundTexture();
+            BuildSimpleBgColor();
+        }
+
+        private void BuildSimpleBgColor()
+        {
+            SpriteRenderer renderer = CreateGameObject("_BgColor");
+            renderer.sprite = LDtkResourcesLoader.LoadDefaultTileSprite();
+            renderer.color = _level.UnityBgColor;
+            
+            _layerSortingOrder.Next();
+            renderer.sortingOrder = _layerSortingOrder.SortingOrderValue;
+
+            ManipulateColorTransform(renderer.transform);
+        }
+
+        private void BuildBackgroundTexture()
+        {
+            //if no path defined, then no background was set.
             if (string.IsNullOrEmpty(_level.BgRelPath))
             {
                 return;
             }
-            
+
             LDtkRelativeGetterLevelBackground getter = new LDtkRelativeGetterLevelBackground();
             _texture = getter.GetRelativeAsset(_level, _importer.assetPath);
-            
+
             if (_texture == null)
             {
                 Debug.LogError("null Sprite");
@@ -47,22 +67,28 @@ namespace LDtkUnity.Editor
                 return;
             }
 
-            SpriteRenderer renderer = CreateGameObject();
+            SpriteRenderer renderer = CreateGameObject("_BgImage");
             renderer.sprite = sprite;
-            
+
             _layerSortingOrder.Next();
             renderer.sortingOrder = _layerSortingOrder.SortingOrderValue;
-            
-            ManipulateTransform(renderer.transform);
-            
+
+            ManipulateImageTransform(renderer.transform);
+
             _importer.AddBackgroundArtifact(sprite);
         }
 
-        private void ManipulateTransform(Transform trans)
+        private void ManipulateColorTransform(Transform trans)
+        {
+            trans.parent = _levelTransform.transform;
+            trans.localPosition = _worldSpaceSize/2;
+            trans.localScale = new Vector3(_worldSpaceSize.x, _worldSpaceSize.y, 1);
+        }
+        private void ManipulateImageTransform(Transform trans)
         {
             trans.parent = _levelTransform.transform;
 
-            Vector2 levelPosition = LDtkCoordConverter.LevelBackgroundPosition(_level.BgPos.UnityTopLeftPx, _level.BgPos.UnityCropRect.height, _importer.PixelsPerUnit, _level.BgPos.UnityScale.y);
+            Vector2 levelPosition = LDtkCoordConverter.LevelBackgroundImagePosition(_level.BgPos.UnityTopLeftPx, _level.BgPos.UnityCropRect.height, _importer.PixelsPerUnit, _level.BgPos.UnityScale.y);
             
             trans.localPosition = levelPosition;
 
@@ -70,9 +96,9 @@ namespace LDtkUnity.Editor
             trans.localScale = new Vector3(scale.x, scale.y, 1);
         }
 
-        private SpriteRenderer CreateGameObject()
+        private SpriteRenderer CreateGameObject(string extraName)
         {
-            GameObject go = new GameObject(_level.Identifier + "_Bg");
+            GameObject go = new GameObject(_level.Identifier + extraName);
             SpriteRenderer renderer = go.AddComponent<SpriteRenderer>();
             return renderer;
         }