+++ Assets/LDtkUnity/Editor/CustomEditor/SceneDrawer/HandleDrawers/LDtkFieldDrawerPoints.cs
@@ -6,22 +6,28 @@ using UnityEngine.Internal;
 
 namespace LDtkUnity.Editor
 {
-    [ExcludeFromDocs]
-    public sealed class LDtkFieldDrawerPoints : ILDtkHandleDrawer
+    public class LDtkFieldDrawerPoints : ILDtkHandleDrawer
     {
         private readonly LDtkFields _fields;
         private readonly string _identifier;
         private readonly EditorDisplayMode _mode;
+        private readonly Vector3 _middleCenter;
 
-        public LDtkFieldDrawerPoints(LDtkFields fields, string identifier, EditorDisplayMode mode)
+        public LDtkFieldDrawerPoints(LDtkFields fields, string identifier, EditorDisplayMode mode, Vector3 middleCenter)
         {
             _fields = fields;
             _identifier = identifier;
             _mode = mode;
+            _middleCenter = middleCenter;
         }
         
         public void OnDrawHandles()
         {
+            if (!LDtkPrefs.ShowFieldPoints)
+            {
+                return;
+            }
+            
             List<Vector2> points = GetConvertedPoints();
             if (points.IsNullOrEmpty())
             {
@@ -77,11 +83,11 @@ namespace LDtkUnity.Editor
 
 
             //if the parsed point was nullable null, then it's going to be negative infinity. Don't draw these null values 
-            convertedRoute.RemoveAll(p => p == Vector2.negativeInfinity);
+            convertedRoute.RemoveAll(HandleAAUtil.IsIllegalPoint);
 
             //round the starting position to the bottom left of the current tile
-            Vector2 pos = _fields.transform.position;
-            pos += (Vector2.one * 0.001f);
+            Vector2 pos = _middleCenter;
+            /*pos += (Vector2.one * 0.001f);
 
             int left = Mathf.FloorToInt(pos.x);
             int right = Mathf.CeilToInt(pos.x);
@@ -89,9 +95,13 @@ namespace LDtkUnity.Editor
 
             int down = Mathf.FloorToInt(pos.y);
             int up = Mathf.CeilToInt(pos.y);
-            pos.y = Mathf.Lerp(down, up, 0.5f);
+            pos.y = Mathf.Lerp(down, up, 0.5f);*/
 
-            convertedRoute.Insert(0, pos);
+            //if we actually have something, then draw our starting point from
+            if (!convertedRoute.IsNullOrEmpty())
+            {
+                convertedRoute.Insert(0, pos);
+            }
             return convertedRoute;
         }
         
@@ -107,7 +117,7 @@ namespace LDtkUnity.Editor
                 pathPoints[i] = points[i];
             }
             
-            HandleAAUtil.DrawAAPath(pathPoints);
+            HandleAAUtil.DrawAAPath(pathPoints, LDtkPrefs.FieldPointsThickness);
         }
         
         /// <summary>
@@ -116,7 +126,7 @@ namespace LDtkUnity.Editor
         private void DrawPathLoop(List<Vector2> points)
         {
             DrawPath(points);
-            HandleAAUtil.DrawAALine(points.First(), points.Last());
+            HandleAAUtil.DrawAALine(points.First(), points.Last(), LDtkPrefs.FieldPointsThickness);
         }
         
         /// <summary>
@@ -129,7 +139,7 @@ namespace LDtkUnity.Editor
             for (int i = 1; i < points.Count; i++)
             {
                 Vector2 nextPointPos = points[i];
-                HandleAAUtil.DrawAALine(pointPos, nextPointPos);
+                HandleAAUtil.DrawAALine(pointPos, nextPointPos, LDtkPrefs.FieldPointsThickness);
             }
         }
 
@@ -138,11 +148,18 @@ namespace LDtkUnity.Editor
         /// </summary>
         private void DrawPoints(List<Vector2> points)
         {
-            Vector3 size = Vector2.one * 0.25f;
+            
+            
+            
+            const float boxUnitSize = 0.2f;
+            
+            float extraThickness = (LDtkPrefs.FieldPointsThickness - 1) * (boxUnitSize/2);
+            
+            Vector3 size = Vector2.one * (boxUnitSize + extraThickness);
             
             foreach (Vector2 point in points)
             {
-                HandleAAUtil.DrawAABox(point, size, fillAlpha: 0);
+                HandleAAUtil.DrawAABox(point, size, fillAlpha: 0, thickness: LDtkPrefs.FieldPointsThickness);
             }
         }