+++ Assets/LDtkUnity/Editor/CustomEditor/SceneDrawer/HandleDrawers/LDtkLevelDrawer.cs
@@ -1,32 +1,35 @@
-﻿using UnityEngine;
+﻿using UnityEditor;
+using UnityEngine;
 
 namespace LDtkUnity.Editor
 {
     public class LDtkLevelDrawer : ILDtkHandleDrawer
     {
+        private readonly GameObject _obj;
         private readonly Vector3 _position;
         private readonly Vector2 _size;
         private readonly string _identifier;
 
-        public LDtkLevelDrawer(Vector3 position, Vector2 size, string identifier)
+        public LDtkLevelDrawer(LDtkComponentLevel level)
         {
-            _position = position;
-            _size = size;
-            _identifier = identifier;
+            _position = level.transform.position;
+            _obj = level.gameObject;
+            _size = level.Size;
+            _identifier = level.Identifier;
         }
 
         public void OnDrawHandles()
         {
             if (LDtkPrefs.ShowLevelIdentifier)
             {
-                HandleUtil.DrawText(_position, _identifier);
+                HandleUtil.DrawText(_identifier, _position, default, () => HandleUtil.SelectIfNotAlreadySelected(_obj));
             }
 
             if (LDtkPrefs.ShowLevelBorder)
             {
                 Vector3 halfSize = _size / 2;
                 Vector3 pos = _position + halfSize;
-                HandleAAUtil.DrawAABox(pos, _size, LDtkPrefs.ShowLevelBorderThickness, 0);
+                HandleAAUtil.DrawAABox(pos, _size, LDtkPrefs.LevelBorderThickness, 0);
             }
         }
     }