+++ Assets/LDtkUnity/Runtime/Components/LDtkComponentLevel.cs
@@ -8,11 +8,18 @@ namespace LDtkUnity
     [ExcludeFromDocs]
     public class LDtkComponentLevel : MonoBehaviour
     {
+        [SerializeField] private string _identifier = string.Empty;
         [SerializeField] private Vector2 _size = Vector2.zero;
         [SerializeField] private Color _bgColor = Color.white;
 
         public Vector2 Size => _size;
         public Color BgColor => _bgColor;
+        public string Identifier => _identifier;
+
+        public void SetIdentifier(string identifier)
+        {
+            _identifier = identifier;
+        }
         
         public void SetSize(Vector2 size)
         {